fix(graph-page): unblock only specified urls
