+++ src/pages/Graph/store/index.ts
@@ -6,3 +6,10 @@ export * as actions from './actions'
 export const store = configureStore({
     reducer: pagesReducer
 })
+
+export const getFaviconsUrls = (): Array<string> => {
+    const { pages } = store.getState()
+    return Object.keys(pages)
+        .map(pageUrl => pages[pageUrl].favIconUrl)
+        .filter(url => !!url) as Array<string>
+}
