+++ src/resourcesLoading/index.ts
@@ -1,19 +1,15 @@
 import { tabs } from '../browser'
 import { installUnblockers } from './install'
-
-export interface ResourcesUrlGetter {
-    (): Array<string>;
-}
+import { ResourcesUrlGetter } from './types'
 
 /** 
  * Favicons can be blocked by CORS polycies in chrome, 
  * if load them from from other sites in extension tab
  * */
 export async function unblockResourcesLoading(getUrls: ResourcesUrlGetter) {
-    // TODO: unblock only given urls
     const { id } = await tabs.getScriptTab()
     if (!id)
         throw new Error("Cannot get id of current tab")
 
-    installUnblockers(id)
+    installUnblockers(id, getUrls)
 }
