+++ src/resourcesLoading/types.ts
@@ -1,0 +1,3 @@
+export interface ResourcesUrlGetter {
+    (): Array<string>;
+}
